added edit subscription
